starting to get library working
