+++ src/test/CMakeLists.txt
@@ -1,5 +1,4 @@
 
-set (CMAKE_CXX_STANDARD 11)
 # https://stackoverflow.com/questions/41251474/how-to-import-zeromq-libraries-in-cmake#41252437
 find_package(PkgConfig)
 ## use pkg-config to get hints for 0mq locations
@@ -15,8 +14,28 @@ find_library(ZeroMQ_LIBRARY
         PATHS ${PC_ZeroMQ_LIBRARY_DIRS}
         )
 
+message("CMAKE_SOURCE_DIR=${CMAKE_SOURCE_DIR}")
+
+add_library(graffsdk SHARED GraffSDK.cpp)
+# set_target_properties(graffsdk PROPERTIES VERSION ${PROJECT_VERSION})
+set_target_properties(graffsdk PROPERTIES PUBLIC_HEADER GraffSDK.h)
+target_include_directories(graffsdk PRIVATE .)
+include(GNUInstallDirs)
+install(TARGETS graffsdk
+        LIBRARY DESTINATION ${CMAKE_INSTALL_LIBDIR}
+        PUBLIC_HEADER DESTINATION ${CMAKE_INSTALL_INCLUDEDIR})
+
 add_executable(feitor-test main.cpp)
 ## add the include directory to our compile directives
 target_include_directories(feitor-test PUBLIC ${ZeroMQ_INCLUDE_DIR})
 ## at the 0mq library to our link directive
 target_link_libraries(feitor-test PUBLIC ${ZeroMQ_LIBRARY})
+# target_link_libraries(feitor-test PUBLIC graffsdk)
+
+# the api testing program
+add_executable(api-test apitest.cpp)
+## add the include directory to our compile directives
+target_include_directories(api-test PUBLIC ${ZeroMQ_INCLUDE_DIR})
+## at the 0mq library to our link directive
+target_link_libraries(api-test PUBLIC ${ZeroMQ_LIBRARY})
+target_link_libraries(api-test PUBLIC graffsdk)
